LPS-75485 Move QuartzTrigger into internal package.
