+++ modules/apps/foundation/portal-scheduler/portal-scheduler-quartz/src/main/java/com/liferay/portal/scheduler/quartz/internal/QuartzTriggerFactory.java
@@ -20,7 +20,6 @@ import com.liferay.portal.kernel.scheduler.TimeUnit;
 import com.liferay.portal.kernel.scheduler.Trigger;
 import com.liferay.portal.kernel.scheduler.TriggerFactory;
 import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.scheduler.quartz.QuartzTrigger;
 
 import java.util.Date;
